Move "Directory" and "DirectoryFile" classes in separate files
