+++ MGDecrypt/DirectoryFile.cs
@@ -1,0 +1,17 @@
+﻿namespace MGDecrypt
+{
+    class DirectoryFile
+    {
+        public bool crypted;
+        public uint offset;
+        public uint length;
+
+        public DirectoryFile(bool crypted, uint offset, uint length)
+        {
+            this.crypted = crypted;
+            this.offset = offset;
+            this.length = length;
+
+        }
+    }
+}
